Further Fixed Links
Turns out in adobe fireworks when you create links from shapes it makes
the hotspot (link) in the shape of a polygon. This may work fine if
exported to html but in my case i am exporting to a PDF where
apparently rectangles are the only acceptable shape. This has caused me
much headache.